refactor: #73 Google Calendar API 호출 코드를 util 클래스로 분리
